requirements.txt
flask==2.3.3
werkzeug==2.3.7
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0 